Allow a window on a secondary display to have focus.
If any window on the default display has focus, then it
gets focus as usual.  If no window on the default display
has focus, then we consider windows on the secondary display.

In the future we will need more elaborate schemes for
managing focus across multiple displays, but this is enough
for testing purposes now.

Bug: 7183618
Change-Id: I21ddb9904eb9e574e42d28743aeca51f4ffebf64

Former-commit-id: 2033763bb18913773eea3ac82bf9dbb053444ac6