Use more reasonable warning thresholds for power on/off.
To encourage vendors to make power on/off as efficient
and responsive as possible, we log some warnings whenever it
is too slow.  The previous thresholds were a little
unreasonable mainly because we wanted to understand
how long the process tends to take even when it's fast.

Raise the warning limit while still being aggressive about
timings.  Ideally we want the screen to turn on/off within
no more than a few frames.

Bug: 7167820
Change-Id: Id28dbf8c91cefa7ae7544b72887104af7aabccff

Former-commit-id: 80cd819a0488d2ce2354e0f11a59e4f01d82ed54