Optical bounds support for LinearLayout, TableLayout and TableRow.
The generic support for optical bounds provided in:

https://googleplex-android-review.googlesource.com/#/c/228269/

did not work correctly for nested LinearLayouts. With this fix
all the layouts in the APIdemos now seem to layout correctly
when the optical bounds mode is enabled.

(There's currently no, user level, way to do this. Instead, the
optical bounds layout mode has to be enabled on a per view basis.)

Change-Id: Id3d1d84445a8a2df277ddfe679c42f2d6cf775f4

Former-commit-id: ad365cc217c989c3b688ae839d6f94ee8889a84e