Revert "Do not use last app rotation as default."
This reverts commit 001e8fae4c0f808f4f341871f6ec0f97d68d6502 [formerly 4bd149ef83c8e3e2ffb61885e71f44df9a9ccfa7]

This change introduced a regression in the device rotation policy.  Suppose you hold the device in landscape orientation then put it flat on a table.  At this time, the accelerometer has no information about which way the device is oriented so the sensor expresses no preference over the orientation.  Now switch to an application that forces portrait.  Then go back to the original app that auto-rotates.  We should remain portrait from this point onwards unless the device is picked up and we have accelerometer information from which to establish a new preferred orientation.

Bug: 7173556
Change-Id: I3684af3369e4ea55042081e19d96773c4d4bbe76

Former-commit-id: 0f862b432e9879790e22cab7dae18caffec0812f