Attempt to fix missing transport control in keyguard
This attempts to fix a bug where the transport control would sometimes be
inoperative.  The problem is that we had two transport control layouts
on some devices because they were being declared in the layout file.

The fix is to only inflate the layout once when KeyguardHostView
is created.

Also removes redundant KeyguardStatusView.

Fixes bug 7254833

Change-Id: Iab84e8326ff745ee57be5177ab2561114c8dc6f0

Former-commit-id: 4e81bd4bcf60a934d4cd589dab6e2bddb88d7356