Rewrite raw insert()s and some raw query()s of moved-to-global keys
The Settings put*() APIs fix up references via the old namespaces,
but the raw insert() interface didn't.  Now it does.  Also, when
possible we fix up direct query() operations on the old namespace
to point to the correct one.  At present that is only done for
query() operations with Uris of the form

    content://secure/adb_enabled

There is no rewriting done on queries of the form

    content://secure WHERE name='adb_enabled'

since the app-supplied WHERE clause can be arbitrarily complex.

Bug 7267568

Change-Id: I5c8cecbea7f5b1da6247a53b1428d3effb0bbca5

Former-commit-id: c221d2be7d2bf57373d43457b18483266f88f9a6