Fix launching home when there are multiple launchers installed
Bug: 7283782

Since ResolverActivity now runs in the correct user, it doesn't need to explicitly
pass in the calling userId. In fact, for the Home case, since ActivityManager is
the one requesting resolution of Home activity, Resolver is running as user 0
instead of the current user.

Removing explicit calls based on the launchingUid fixes the problem.

Change-Id: I4676c155aafaa0dd5cefb1acefa0e7e9c1e20f25

Former-commit-id: 203a2f464cef0d873ba8a6193e288fd25b6dc907