Ensure we send at least one wakeup/gotosleep transition.
This fixes an issue where the device would not lock immediately
when the user quickly pressed power off / power on even if
configured to do so.  We were suppressing the screen off
and wake up broadcast in this case.

Now we make sure to always send at least one broadcast to
indicate the transition.  We still collapse back-to-back
full cycle transitions though so as not to end up enqueuing
useless broadcasts.

Bug: 7061116
Change-Id: I7211c5fd963c271c2b0aceb4d2f746063c629079

Former-commit-id: 543083593430208af87e51b94cdbbb51cbe2d7b5