Lubricate a sticky notification panel.
If the panel was left open when the screen was turned off,
in some cases it might get stuck in an "open" state (the
panel's expanded height would be nonzero) although the
status bar window is in fact fully collapsed (due to
makeExpandedInvisible). The next time the user would go to
open the panel, things would be in an inconsistent state and
the panel wouldn't come down (on phones, the settings panel
would be attempted, but still nothing would happen).

This was easiest to reproduce on the keyguard (turn on
screen, pull down panel, turn off screen).

Bug: 7260868
Change-Id: Iec0000ba020e5a519eb5b4d42ac273b6689a18bd

Former-commit-id: fc82b56d1b0831631c880f2e6bf7bdc506994872