Keep quick settings brightness icon/dialog level in sync.
We now listen for brightness changes and update the icon appropriately.
Note there are still only two icon states in QS (auto-on and auto-off).

We also listen for user switches, since the brightness setting is per-user.

The manual brightness level is now kept in sync with level set in Settings.

This change is tightly scoped to brightness only.  Any other
QS bugs should be triaged separately.

Bug:7283805
Change-Id: Ibaf8fe94e98f65cc02efe7fcddf4f38370fdb09f

Former-commit-id: 48f37ec8c98a6281dbf067d1740a3bc4953ed26f