Quick settings user display and switching
Load user information for the current user and display in quick
settings tile. Name is pulled from the contacts DB from the "Me"
profile, which will override the raw name on the system user if
needed.

Open the user switcher on the lock screen if this tile is clicked and
there is more than one user registered on the system. If there is only
one user, show the "Me" quick contact card.

Darken the background protection behind the user's name on the QS tile
layout so that it's readable against very light profile pic
backgrounds.

Bug 7175023
Bug 7257997

Change-Id: Ia1e7dd7af72dbd49113a827f9228e0a32e20a0dc

Former-commit-id: 6f2a3d298205522f2e799800dace0e6c345dd4d3