Suppressing FUL after user switch (fix b/7316467)
When switching users, Face Unlock starts in onResume().  However,
there is no signal to indicate when the user actually sees their
unlock screen.  This means Face Unlock could be running unseen, timing
out soon after it becomes visible, or letting the user in before they
see the preview.

This fix simply suppresses Face Unlock immediately after switching
users.  This is not the ideal behavior, but there is no easy way to
make Face Unlock start only after the unlock screen becomes visible.
When the user changes screens it becomes unsuppressed, so if they go
back to the multi-select widget screen or login, Face Unlock works as
expected and is only suppressed again when the user is switched.

Change-Id: I80a302b0aefc1dee3c2dc77557978cbe062de435

Former-commit-id: cc4104fd71225ed092f16b11882d9a10084e34ac