Don't enable input dispatch until the screen is visible.
When we defer making the screen visible (waiting for the lock screen
to be ready) the screen may actually be on but covered by a black
surface.  We need to make sure to ignore any touches on the screen
during this time until the black surface is about to be removed.

Bug: 7318962
Change-Id: I50eb7dcf05295cd276925625240996c4b80c5fe2

Former-commit-id: 4fc452795f8f1ad1a9c26720037836e9d5d0db2f