Reducing WebView navigation timeout for accessibility.
Reducing the amount of time that WebView will wait before treating
a navigation attempt as failed. This is needed to recover from
cases where the user is attempting to linearly navigate a WebView
that is immediately switched out right as it starts to read.
A good example of this problem is AdMob's WebViews which are
continuously being swapped.

Change-Id: Ib9b4a9b99a35d42920e34e0acbe8ab45e1d47871

Former-commit-id: 2203cf3f725220e6dc930ee77dbcde4ff7d885b6