Fix keyguard dimensions; IME state for password view; layouts
Fix a bug where ordering during inflation caused us not to set a
keyguard callback early enough to properly modify window flags. Add a
gross hack to work around this for now.

Fix status layouts to scale a bit more gracefully in the presence of
an IME.

Fix password layouts to scale more gracefully in the presence of an
IME

Bug 7343312
Bug 7341795
Bug 7342963
Bug 7343089

Change-Id: Ifb2c06defef11e8f7f9d0e09855632ed491bb31c

Former-commit-id: 70bc9f24ce6098a8c82245ac1ce7721d354ecd3d