Fix for bug 7344916 text view flickers when you hit enter
The flickering was caused by trying to scroll to the cursor position
while the view was in an inconsistent state (text updated to change the
number of lines, but layout not done yet). This patch defers the actual
setting of the cursor until layout is done, when layout is pending.

Change-Id: I8ed3a402beb8058ac7a7f3935afeb946a23308ab

Former-commit-id: f5c1a8798f3a2e4196473fb1328159bd8297f178