Update Arabic fonts to remove Rial ligature
This patch fixes bug 7309211 [ar][fa] Riyal ligature mistakenly on in
Android Arabic fonts

It also fixes a number of other problems in the normal (non-system ui)
versions of the Naskh fonts, similar to the changes applied for the fix
to bug 6872613, and makes those fonts more consistent with the system ui
version.

3. Fixed a wrong GSUB substitution for Urdu Seven
4. Fixed shaping of U+06C1 HEH GOAL
5. Fixed shaping of U+06BA NOON GHUNNA
6. Removed the Rial ligature from the GSUB table
7. Tightened the Alef-Lam-Lam-Heh ligature in the GSUB table which was
getting triggered too frequently

Thanks to Roozbeh for the updated font data.

Change-Id: I29aecba2ea62853a8ed3181c023c62a09e800def

Former-commit-id: cd31db60ee6df6dd84bce9925e96b6cccbdf522d