Layoutlib: Read and close XML files as soon as possible.
Because passing an InputStream to KXML does not close the
stream after the file has been parsed, the files are staying
locked on windows until the gc and finalizers are run.

This change preload the XML files and close their stream,
and then pass the content in a stream to the parser.

Change-Id: Iabe27989dc616ec9e7de88e52b1ec3af9f007f7c

Former-commit-id: c48b0b940d632b9862ff6aa32665e37fe9e50845