Fix nav bar glitch when quickly turning screen off then on.
Keyguard now disables all navbar buttons when turning screen off
instead of only recents.

Navbar is told about screen off/on and disables the layout transition
animation on screen off - if the user quickly turns the screen back
on they do not see the end of the fade out animation.

Bug:7377776
Change-Id: I96a050d58c6c9ce8537f50408655006fadbfcd4f

Former-commit-id: 1bbd49d72eee001137b6d6e6ab3f353fe2c0433c