Reset ExpandHelper when the panel is collapsed.
Fixes a rather unpleasant bug in which the ExpandHelper
could get locked in "expanding" mode if the panel was closed
(for example, with the back button) while you were in the
middle of an expand gesture. In this situation ExpandHelper
would hungrily eat all future touch events destined for the
notification panel, making it impossible to click or even
clear any notifications.

Bug: 7330828
Change-Id: I9c493db5e8fd8ef1aca53f77820780d60fa4e5a7

Former-commit-id: ac47ff70c322614ff2ca9ad82fe41338daf55877