Handling popups correctly for face unlock
Added a keyguardVisibilityChanged callback to indicate when the
keyguard is covered by something (alarm, phone, etc).  Face unlock
doesn't start if it's not visible and it restarts appropriately when
it gains visibility again.  This mostly fixes b/7362774 and b/7376783,
but there's some jank when face unlock goes away (if it was interrupted
while running).

Change-Id: I7bdf8ef6af687acaa7152d8e61bfdcc1979259ff

Former-commit-id: f6d0fc190c4b257ee4f2f1ae8a59519d3a21ce6d