FUL uses square aspect ratio for all layouts fixes b/7426399
When the referenced bug was filed, face unlock was only showing as a
tiny sliver on tablet layouts.  Due to other changes in the lockscreen
branch, it was no longer a sliver, but was still an incorrect layout,
with a width much greater than height.  This change makes face unlock
square for all layouts.

The face unlock RelativeLayout was replaced with a custom
FaceUnlockView derived from RelativeLayout.  The new view forces a
square layout using the same technique used by LockPatternView.

Note that there is still a bug where the pattern view covers the widget
area on portrait tablet layouts.  The face unlock view has this same
issue, but may resolve itself when the pattern bug is fixed.

Also note there are two other Face Unlock tablet bugs that existed
before this change (and therefore are not caused by this change):
 - flash during transition from spotlight animation to camera preview
 - PIN backup is upper-left instead of centered

Change-Id: I550eccfa3924f230a9dc43f0a9b59f1ea55a9273

Former-commit-id: 68d257d788f5154922cd605ab8079a5c0815dffc