Fix for bug 7400445 regression in background drawable
The opacity calculation for a gradient drawable of shape must take
rounded corners into account - if the corner radius is nonzero, then the
shape is translucent rather than opaque. Previously the code always
assumed that such rectangles were fully opaque, which led to the
background (visible behind the rectangle) not getting drawn.

This patch simply checks for corner radius in addition to shape and
computes opacity as translucent in the nonzero case.

Change-Id: Iaf4d24abc6ecf49f85c82972b8f998700c83295e

Former-commit-id: 39824f685c33f8aabbe9526b4cc79b7411e8b37f