Clearing connected message in stop fixes 7401152
We can stop before the service connected message is processed, in which case
we should clear the message.  Otherwise it can cause us to start up the UI
before the camera is open, causing a crash.

I was unable to reproduce this issue, but it theoretically will happen, and
the bug report shows all the symptoms of this series of events.

Change-Id: I60d295883b08ac58ccf0f3fc6d152b8b75fe9b18

Former-commit-id: 025fb93e3e5518b7db5fec049d285cfb2d28a074