Removing callbacks in onDetachedFromWindow, fixes b/7465673
In certain cases when lockscreen is constructed in landscape and then
switches to portrait on the phone, it inflates new views for
KeyguardHostView.  The old views never call onPause() in KeyguardFaceUnlockView
so it was still getting callbacks and starting Face Unlock multiple
times.  By removing the callbacks in onDetachedFromWindow() there won't be
multiple instances of KeyguardFaceUnlockView receiving callbacks and
starting face unlock.

Change-Id: I00d770283e83be2f4b76e00340123789a8075091

Former-commit-id: cade0049a729ac63832f45befa91cec3b7ce88f5