Use a consistent policy for filtering wake keys.
Previously wake keys were filtered differently depending on whether
a keyguard was showing.  If the user disables the keyguard then
no filtering was applied, which means that the volume key may
wake your device while in your pocket.

This change ensures that we use the same policy consistently
regardless of whether keyguard is showing.  The behavior is
otherwise the same.

Removed the "Locked" suffix on a method that was actually
being called without a lock held and which in fact does not
require it.

Bug: 7481025
Change-Id: I704c71ca009bc5437f349f858b9de7c77ea73e4b

Former-commit-id: 1c2e4948a1f0e803171389427212321973ea66b8