Fix "keyguard unresponsive" bug
This fixes a bug where the code would pop up SearchPanelView and never
close it.  The result is that the overlapping window would absorb all input
events and effectively prevent the user from entering their pin/pattern/password.

The fix is to replace ImageView with KeyButtonView to make the behavior consistent
with how it's handle in home screen.

Fixes bug 7520075

Change-Id: I461bf2844c023fd7a2dffca1177cc5951f30dba0

Former-commit-id: c120b6e99ab9982b046772e8d41cf4f27e9f1026