Revert "NumberPicker should adjust min and max when displayed values are set." (a.k.a. Santa is back)
This reverted change was adjusting the min and max values for the NumberPicker
which is not desirable since it changes behavior and it will be possible for
an app that works on the current platform to crash on an older one. Also the
adjustment was not implemented correctly.

Updated the documentation to clarify the reltionship between the min value,
max value, and the displayed values array.

Bug:7518172

This reverts commit b0817aa68690b0a0a2d8148ec8e570cd8c150dbc [formerly a1410e6789ce72bc423793315a51aea8b6bad6c7]

Change-Id: I109f1b1f54c1e609941243cabab9241871b6b12b

Former-commit-id: 7018cfdc05dc6135949806749ff5c370dce09ced